#885 Moved helper classes to a separate sub-package.
