+++ platform-dao/src/main/java/ua/com/fielden/platform/companion/AbstractEntityReader.java
@@ -3,7 +3,7 @@ package ua.com.fielden.platform.companion;
 import static java.lang.String.format;
 import static java.util.Collections.emptyMap;
 import static java.util.stream.Collectors.toList;
-import static ua.com.fielden.platform.companion.KeyConditionBuilder.createQueryByKey;
+import static ua.com.fielden.platform.companion.helper.KeyConditionBuilder.createQueryByKey;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.fetchAggregates;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.from;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.orderBy;
