+++ platform-dao/src/main/java/ua/com/fielden/platform/companion/PersistentEntitySaver.java
@@ -3,10 +3,10 @@ package ua.com.fielden.platform.companion;
 import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
 import static org.hibernate.LockOptions.UPGRADE;
-import static ua.com.fielden.platform.companion.ActivatableEntityRetrospectionHelper.addToResultIfApplicableFromActivatablePerspective;
-import static ua.com.fielden.platform.companion.ActivatableEntityRetrospectionHelper.collectActivatableNotDirtyProperties;
-import static ua.com.fielden.platform.companion.ActivatableEntityRetrospectionHelper.isNotSpecialActivatableToBeSkipped;
-import static ua.com.fielden.platform.companion.KeyConditionBuilder.createQueryByKey;
+import static ua.com.fielden.platform.companion.helper.ActivatableEntityRetrospectionHelper.addToResultIfApplicableFromActivatablePerspective;
+import static ua.com.fielden.platform.companion.helper.ActivatableEntityRetrospectionHelper.collectActivatableNotDirtyProperties;
+import static ua.com.fielden.platform.companion.helper.ActivatableEntityRetrospectionHelper.isNotSpecialActivatableToBeSkipped;
+import static ua.com.fielden.platform.companion.helper.KeyConditionBuilder.createQueryByKey;
 import static ua.com.fielden.platform.entity.ActivatableAbstractEntity.ACTIVE;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.from;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.select;
