+++ platform-dao/src/main/java/ua/com/fielden/platform/security/dao/UserAndRoleAssociationDao.java
@@ -1,6 +1,6 @@
 package ua.com.fielden.platform.security.dao;
 
-import static ua.com.fielden.platform.companion.KeyConditionBuilder.createQueryByKeyFor;
+import static ua.com.fielden.platform.companion.helper.KeyConditionBuilder.createQueryByKeyFor;
 
 import java.util.Set;
